+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- New `ConversationActivity` items are now correctly handled when the chat is open.
+
 
 ## [1.0.0-alpha04] - 2022-06-01